+++ Long/quicksortLST.h
@@ -5,7 +5,7 @@
 #ifndef DISSERTATION_QUICKSORTST_H
 #define DISSERTATION_QUICKSORTST_H
 typedef unsigned long long UINT64;
-void quickSortLST(UINT64 arr[], int left, int right);
+void quickSortLST(UINT64 arr[], UINT64 left, UINT64 right);
 
 
 #endif //DISSERTATION_QUICKSORTST_H
